+++ lib/tlTimeline/AudioInline.h
@@ -6,6 +6,27 @@ namespace tl
 {
     namespace timeline
     {
+        inline bool AudioLayer::operator == (const AudioLayer& other) const
+        {
+            return audio == other.audio;
+        }
+
+        inline bool AudioLayer::operator != (const AudioLayer& other) const
+        {
+            return !(*this == other);
+        }
+
+        inline bool AudioData::operator == (const AudioData& other) const
+        {
+            return seconds == other.seconds &&
+                layers == other.layers;
+        }
+
+        inline bool AudioData::operator != (const AudioData& other) const
+        {
+            return !(*this == other);
+        }
+
         inline bool isTimeEqual(const AudioData& a, const AudioData& b)
         {
             return a.seconds == b.seconds;
