+++ lib/tlTimeline/CMakeLists.txt
@@ -1,25 +1,31 @@
 set(HEADERS
 	Audio.h
 	AudioInline.h
+	CompareOptions.h
+	CompareOptionsInline.h
+	DisplayOptions.h
+	DisplayOptionsInline.h
+	ImageOptions.h
+	ImageOptionsInline.h
 	IRender.h
-	RenderOptions.h
-	RenderOptionsInline.h
     Timeline.h
     TimelinePlayer.h
     TimelinePlayerInline.h
+	Transition.h
 	Util.h
 	UtilInline.h
 	Video.h
 	VideoInline.h)
 
 set(SOURCE
-	Audio.cpp
+	CompareOptions.cpp
+	DisplayOptions.cpp
+	ImageOptions.cpp
 	IRender.cpp
-	RenderOptions.cpp
     Timeline.cpp
     TimelinePlayer.cpp
-	Util.cpp
-	Video.cpp)
+	Transition.cpp
+	Util.cpp)
 
 add_library(tlTimeline ${HEADERS} ${SOURCE})
 target_link_libraries(tlTimeline tlIO)
